bot/requirements.txt
psycopg2
slack_bolt
SQLAlchemy
python-dotenv